Move Alerts to layout folder and update imports
