+++ client/src/components/CartItems.js
@@ -4,7 +4,7 @@ import Modal from "../components/Modal"
 import ShippingForm from "./ShippingForm.js"
 import AppContext from "../context/AppContext";
 import AlertContext from "../context/alertContext";
-import Alerts  from "../components/Alerts"
+import Alerts  from "../components/layout/Alerts"
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 
 const CartItems = () => {
