+++ client/src/pages/Contact.js
@@ -2,7 +2,7 @@ import React, {useContext, useState} from 'react'
 import "../styles/contact.css"
 import AlertContext from "../context/alertContext";
 import axios from "axios"
-import Alerts from "../components/Alerts"
+import Alerts from "../components/layout/Alerts"
 import PageHeader from "../components/layout/PageHeader"
 
 const Contact = () => {
